app/requirements.txt
pandas==1.5.3
psycopg2-binary==2.9.5